[fix] Fixed multiple physics bug related to big delta time (for cases when game lags too much or paused using debugger)

+++ core/ecs/Constants.h
@@ -5,8 +5,8 @@
 #pragma once
 
 namespace constants {
-    const int DEFAULT_X_SPEED = 10;
-    const int DEFAULT_JUMP_FORCE = 20;
-    const int GRAVITY_ACCELERATION = 1;
-    const int MAX_FALL_SPEED = 50;
+    const float DEFAULT_X_SPEED = 10;
+    const float DEFAULT_JUMP_FORCE = 20;
+    const float GRAVITY_ACCELERATION = 1;
+    const float MAX_FALL_SPEED = 50;
 }
